migrate k8s hooks to SDK
