+++ frontend/packages/console-dynamic-plugin-sdk/src/app/k8s/reducers/k8sSelector.ts
@@ -1,0 +1,1 @@
+export const getReduxIdPayload = (state, reduxId) => state.sdkK8s.get(reduxId);
