+++ frontend/packages/console-dynamic-plugin-sdk/src/utils/k8s/hooks/k8s-watch-types.ts
@@ -1,0 +1,20 @@
+import { Dispatch } from 'react-redux';
+import { K8sModel } from '../../../api/common-types';
+import { WatchK8sResource } from '../../../extensions/console-types';
+
+export type RootState = any;
+
+export type GetIDAndDispatch = (
+  resource: WatchK8sResource,
+  k8sModel: K8sModel,
+) => { id: string; dispatch: (dispatch: Dispatch, getState: () => RootState) => void };
+
+export type Query = { [key: string]: any };
+
+export type MakeQuery = (
+  namespace: string,
+  labelSelector?: any,
+  fieldSelector?: any,
+  name?: string,
+  limit?: number,
+) => Query;
