+++ frontend/packages/console-dynamic-plugin-sdk/src/utils/k8s/hooks/usePrevious.ts
@@ -1,0 +1,10 @@
+import * as React from 'react';
+
+export const usePrevious = <P = any>(value: P, deps?: any[]): P => {
+  const ref = React.useRef<P>();
+  React.useEffect(() => {
+    ref.current = value;
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, deps || [value]);
+  return ref.current;
+};
